+++ lib/custom_lints_template.dart
@@ -1,5 +1,5 @@
 import 'package:custom_lint_builder/custom_lint_builder.dart';
-import 'package:custom_lints_template/src/lints/avoid_hardcoded_strings.dart';
+import 'package:custom_lints_template/src/lints/avoid_hardcoded_strings/avoid_hardcoded_strings_rule.dart';
 
 PluginBase createPlugin() => _CustomLintTemplatePlugin();
