+++ lib/src/lints/avoid_hardcoded_strings/fixes/avoid_hardcoded_strings_fix.dart
