+++ lib/src/lints/avoid_hardcoded_strings/models/avoid_hardcoded_strings.dart
@@ -1,6 +1,6 @@
 import 'package:analyzer/error/error.dart';
-import 'package:custom_lints_template/src/options/hooks/error_severity_hook.dart';
-import 'package:custom_lints_template/src/options/hooks/map_or_list_hook.dart';
+import 'package:custom_lints_template/src/models/hooks/error_severity_hook.dart';
+import 'package:custom_lints_template/src/models/hooks/map_or_list_hook.dart';
 import 'package:dart_mappable/dart_mappable.dart';
 
 part 'avoid_hardcoded_strings.mapper.dart';
