+++ lib/src/lints/avoid_hardcoded_strings/models/avoid_hardcoded_strings.mapper.dart
