+++ lib/src/lints/custom_lint_example/custom_lint_example.dart
@@ -1,9 +1,9 @@
 import 'package:analyzer/error/listener.dart';
 import 'package:custom_lint_builder/custom_lint_builder.dart';
-import 'package:custom_lints_template/src/lints/fixes/fix_example.dart';
+import 'package:custom_lints_template/src/lints/custom_lint_example/fixes/fix_example.dart';
+import 'package:custom_lints_template/src/lints/custom_lint_example/models/custom_lint_example.dart';
 import 'package:custom_lints_template/src/models/options_lint_rule.dart';
 import 'package:custom_lints_template/src/models/rule_config.dart';
-import 'package:custom_lints_template/src/options/custom_lint_example.dart';
 import 'package:custom_lints_template/src/utils/extensions.dart';
 import 'package:custom_lints_template/src/utils/path_utils.dart';
 
@@ -16,13 +16,11 @@ class CustomLintExampleRule extends OptionsLintRule<CustomLintExampleOption> {
           configs: configs,
           name: lintName,
           paramsParser: CustomLintExampleOption.fromMap,
-          problemMessage: (_) =>
-              'Avoid hardcoding strings. Use a localization package or append ".hardcoded" to the string to suppress this message.',
+          problemMessage: (_) => 'This is the description of our custom lint',
         ));
 
-  /// The [LintCode] of this lint rule that represents
-  /// the error whether we use bad formatted double literals.
-  static const String lintName = 'avoid_hardcoded_strings';
+  /// The [LintCode] of this lint rule
+  static const String lintName = 'custom_lint_example';
 
   @override
   Future<void> run(
