+++ lib/src/lints/custom_lint_example/models/custom_lint_example.mapper.dart
