+++ lib/src/models/hooks/error_severity_hook.dart
